Finish nesting html elements and begin CSS styling
Wrapping the text under each grid outline is not working. I think I am
flexwrapping the wrong divs. Also, I can't get the credit at the end of
the quote text to flex-end. It is ending outside of the text margin.

A constant problem seems to be distinguishing the div which I use to
style the background and the div which contains my content. They
conflict and it makes the flexbox do things I am not expecting.